+++ src/components/OldMessages/index.tsx
@@ -1,0 +1,30 @@
+import { GridItemEight, GridItemFour, GridLayout } from '@components/UI/GridLayout';
+import MetaTags from '@components/utils/MetaTags';
+import isFeatureEnabled from '@lib/isFeatureEnabled';
+import { NextPage } from 'next';
+import React from 'react';
+import { APP_NAME } from 'src/constants';
+import Custom404 from 'src/pages/404';
+import { useAppStore } from 'src/store/app';
+
+const Messages: NextPage = () => {
+  const currentProfile = useAppStore((state) => state.currentProfile);
+
+  if (!isFeatureEnabled('messages', currentProfile?.id)) {
+    return <Custom404 />;
+  }
+
+  return (
+    <GridLayout>
+      <MetaTags title={`Messages • ${APP_NAME}`} />
+      <GridItemFour>
+        <div>gm</div>
+      </GridItemFour>
+      <GridItemEight>
+        <div>gm</div>
+      </GridItemEight>
+    </GridLayout>
+  );
+};
+
+export default Messages;
