+++ src/components/Shared/Navbar/index.tsx
@@ -51,7 +51,7 @@ const Navbar: FC = () => {
       <>
         <NavItem url="/" name="Home" current={pathname == '/'} />
         <NavItem url="/explore" name="Explore" current={pathname == '/explore'} />
-        <NavItem url="/conversations" name="Messages" current={pathname == '/conversations'} />
+        <NavItem url="/messages" name="Messages" current={pathname == '/messages'} />
         <MoreNavItems />
       </>
     );
