--- src/pages/conversations/index.tsx
@@ -1,3 +1,0 @@
-import Conversations from '@components/Conversations';
-
-export default Conversations;
