+++ src/pages/messages/index.tsx
@@ -1,0 +1,3 @@
+import Messages from '@components/Messages';
+
+export default Messages;
